+++ src/main/java/com/farmdora/farmdoraactivity/common/response/SuccessMessage.java
@@ -15,7 +15,8 @@ public enum SuccessMessage {
     SEARCH_TOTALUSERS_SUCCESS("사이트 가입자 수 조회에 성공하였습니다."),
     SEARCH_USERDASHBOARDINFO_SUCCESS("내 정보 조회에 성공하였습니다."),
     SEARCH_ORDERSTATUSINFO_SUCCESS("주문 현황 조회에 성공하였습니다."),
-    SEARCH_WISHLIST_SUCCESS("찜 리스트 조회에 성공하였습니다.");
+    SEARCH_WISHLIST_SUCCESS("찜 리스트 조회에 성공하였습니다."),
+    SEARCH_DELETEWISH_SUCCESS("찜 항목 삭제에 성공하였습니다.");
