+++ src/main/java/com/farmdora/farmdoraactivity/user/repository/UserDashboardRepository.java
@@ -44,19 +44,4 @@ public interface UserDashboardRepository extends JpaRepository<User, Integer> {
             "JOIN SaleFile sf ON s.id = sf.sale.id AND sf.isMain = false " +
             "WHERE l.user.userId = :userId")
     List<Object[]> findWishPreviewByUserId(@Param("userId") Integer userId, Pageable pageable);
-
-    @Query("SELECT l.id, s.id, s.title, o.name, o.price, sf.saveFile, " +
-            "ROUND(AVG(re.score), 1), COUNT(DISTINCT re.id) " +
-            "FROM Like l " +
-            "JOIN l.sale s ON l.sale = s " +
-            "JOIN Option o ON o.sale = s " +
-            "JOIN SaleFile sf ON sf.sale = s AND sf.isMain = false " +
-            "LEFT JOIN Review re ON re.sale = s " +
-            "WHERE l.user.userId = :userId " +
-            "AND o.id = (SELECT MIN(o2.id) FROM Option o2 WHERE o2.sale = s) " +
-            "GROUP BY s.id, s.title, o.name, o.price, sf.saveFile " +
-            "ORDER BY l.id DESC")
-    List<Object[]> findWishlistByUserId(@Param("userId") Integer userId);
-
-
 }
