+++ src/main/java/com/farmdora/farmdoraactivity/user/service/UserDashboardService.java
@@ -15,5 +15,4 @@ public interface UserDashboardService {
 
     List<WishPreviewDTO> getWishPreviewByUserId(Integer userId);
 
-    List<WishlistDTO> getWishlistByUserId(Integer userId);
 }
