+++ src/main/java/com/farmdora/farmdoraactivity/user/service/UserWishlistService.java
@@ -1,0 +1,13 @@
+package com.farmdora.farmdoraactivity.user.service;
+
+import com.farmdora.farmdoraactivity.user.dto.WishlistDTO;
+
+import java.util.List;
+
+public interface UserWishlistService {
+
+    List<WishlistDTO> getWishlistByUserId(Integer userId);
+
+    void deleteSelectedWishlist(List<Integer> likeIds);
+
+}
